requirements.txt
bs4
mysql-connector-python